+++ include/mappoint.h
@@ -12,8 +12,8 @@ namespace myslam{
             unsigned long                       id_;        // ID
             static unsigned long                factory_id_;    // factory id
             bool                                good_;      // wheter a good point 
-            Vector3f                            pos_;       // Position in world
-            Vector3f                            norm_;      // Normal of viewing direction 
+            Vector3d                            pos_;       // Position in world
+            Vector3d                            norm_;      // Normal of viewing direction 
             cv::Mat                             descriptor_; // Descriptor for matching 
             std::list<Frame*>                   observed_frames_;   // key-frames that can observe this point 
             int                                 matched_times_;     // being an inliner in pose estimation
@@ -22,20 +22,20 @@ namespace myslam{
             MapPoint();
             MapPoint( 
                 unsigned long id, 
-                const Vector3f& position, 
-                const Vector3f& norm, 
+                const Vector3d& position, 
+                const Vector3d& norm, 
                 Frame* frame=nullptr, 
                 const cv::Mat& descriptor=cv::Mat() 
             );
             
-            inline cv::Point3f getPositionCV() const {
-                return cv::Point3f( pos_(0,0), pos_(1,0), pos_(2,0) );
+            inline cv::Point3d getPositionCV() const {
+                return cv::Point3d( pos_(0,0), pos_(1,0), pos_(2,0) );
             }
             
             static MapPoint::Ptr createMapPoint();
             static MapPoint::Ptr createMapPoint( 
-            const Vector3f& pos_world, 
-            const Vector3f& norm_,
+            const Vector3d& pos_world, 
+            const Vector3d& norm_,
             const cv::Mat& descriptor,
             Frame* frame );
     };
