+++ src/frame.cpp
@@ -43,16 +43,16 @@ void Frame::setPose ( const SE3& T_c_w ){
     T_c_w_ = T_c_w;
 }
 
-Vector3f Frame::getCamCenter() const{
+Vector3d Frame::getCamCenter() const{
     return T_c_w_.inverse().translation();
 }
 
-bool Frame::isInFrame ( const Vector3f& pt_world )
+bool Frame::isInFrame ( const Vector3d& pt_world )
 {
-    Vector3f p_cam = camera_->world2camera( pt_world, T_c_w_ );
+    Vector3d p_cam = camera_->world2camera( pt_world, T_c_w_ );
     // cout<<"P_cam = "<<p_cam.transpose()<<endl;
     if ( p_cam(2,0)<0 ) return false;
-    Vector2f pixel = camera_->world2pixel( pt_world, T_c_w_ );
+    Vector2d pixel = camera_->world2pixel( pt_world, T_c_w_ );
     // cout<<"P_pixel = "<<pixel.transpose()<<endl<<endl;
     return pixel(0,0)>0 && pixel(1,0)>0 
         && pixel(0,0)<color_.cols
